# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Nabra Space API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /generations:
    get:
      operationId: listGenerations
      tags: [generation]
      summary: List user generation history
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of generations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Generation"
  /generations/{id}:
    get:
      operationId: getGeneration
      tags: [generation]
      summary: Get generation status by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Generation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Generation"
        "404":
          description: Not found
...
  /affiliate/attribute:
    post:
      operationId: attributeReferral
      tags: [affiliate]
      summary: Attribute referral on signup
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttributeInput"
      responses:
        "200":
          description: Attribution result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttributeResult"
  /affiliate/stats:
    get:
      operationId: getAffiliateStats
      tags: [affiliate]
      summary: Get current user affiliate stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Affiliate stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AffiliateStats"
  /affiliate/click:
    post:
      operationId: trackAffiliateClick
      tags: [affiliate]
      summary: Track affiliate link click
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClickInput"
      responses:
        "200":
          description: Click tracked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClickResult"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    GenerationBlock:
      type: object
      required: [tone, text]
      properties:
        tone: { type: string }
        text: { type: string }
    Generation:
      type: object
      required: [id, user_id, status, created_at]
      properties:
        id: { type: string }
        user_id: { type: string }
        text_input: { type: ["string", "null"] }
        character_count: { type: ["integer", "null"] }
        blocks_json:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/GenerationBlock"
        audio_url: { type: ["string", "null"] }
        status: { type: string }
        created_at: { type: string }
...
    AffiliateStats:
      type: object
      required: [user_id, referral_code, commission_rate, total_clicks, total_signups, total_conversions, pending_balance, available_balance, status]
      properties:
        user_id: { type: string }
        referral_code: { type: string }
        commission_rate: { type: number }
        total_clicks: { type: integer }
        total_signups: { type: integer }
        total_conversions: { type: integer }
        pending_balance: { type: number }
        available_balance: { type: number }
        status: { type: string }
...
    AttributeInput:
      type: object
      required: [referral_code]
      properties:
        referral_code: { type: string }
    AttributeResult:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
        message: { type: ["string", "null"] }
    ClickInput:
      type: object
      required: [referral_code]
      properties:
        referral_code: { type: string }
    ClickResult:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
...