# lib/api-zod/src/generated/types/attributeResult.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Nabra Space API
 * OpenAPI spec version: 0.1.0
 */

export interface AttributeResult {
  success: boolean;
  /** @nullable */
  message?: string | null;
}